requirements.txt
alabaster==0.7.12
antlr4-python3-runtime==4.9.2
argparse==1.4.0
asgiref==3.4.1
attrs==21.2.0
babel==2.9.1
bcp47==0.0.4
beautifulsoup4==4.10.0
bidict==0.21.3
bioregistry==0.3.30
cachetools==4.2.4
certifi==2021.10.8
cfgraph==0.2.1
charset-normalizer==2.0.7; python_version >= '3'
click==7.1.2
commonmark==0.9.1
curieutil==0.0.5
decorator==5.1.0
defusedxml==0.7.1
deprecated==1.2.13
diskcache==5.2.1
docopt==0.6.2
docutils==0.17.1
et-xmlfile==1.1.0
frozendict==2.0.6
funowl==0.1.9
ghp-import==2.0.2
googlemaps==4.5.3
graphviz==0.17
greenlet==1.1.2; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
h11==0.12.0
hbreader==0.9.1
httptools==0.2.0
idna==3.3; python_version >= '3'
imagesize==1.2.0
importlib-metadata==4.8.1
inflect==5.3.0
iniconfig==1.1.1
isodate==0.6.0
jinja2==3.0.2
json-flattener==0.1.7
jsonasobj2==1.0.4
jsonasobj==2.0.2.dev4
jsonpatch==1.32
jsonpath-ng==1.5.3
jsonpointer==2.1
jsonschema==4.1.2
linkml-model==1.0.5
linkml-runtime-api==0.0.4
linkml-runtime==1.1.3
linkml==1.1.11
lxml==4.6.3
markdown-it-py==1.1.0
markdown==3.3.4
markupsafe==2.0.1
mdit-py-plugins==0.2.8
mergedeep==1.3.4
mkdocs==1.2.3
more-click==0.0.6
more-itertools==8.10.0
myst-parser==0.15.2
nmdc-schema==2021.10.20rc1
num2words==0.5.10
numpy==1.21.3; platform_machine != 'aarch64' and platform_machine != 'arm64' and python_version < '3.10'
openpyxl==3.0.9
packaging==21.0
pandas==1.3.4
parse==1.19.0
pint==0.17
pluggy==1.0.0
ply==3.11
prefixcommons==0.1.9
prologterms==0.0.6
py==1.10.0
pydantic==1.8.2
pygments==2.10.0
pyjsg==0.11.6
pyldmod==2.0.5
pyparsing==3.0.0rc2
pyrsistent==0.18.0
pyshex==0.7.20
pyshexc==0.8.3
pystow==0.2.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pyyaml-env-tag==0.1
pyyaml==5.4.1
quantulum3==0.7.9
rdflib-jsonld==0.6.1
rdflib-pyldmod-compat==0.1.2
rdflib==5.0.0
recommonmark==0.7.1
requests-ftp==0.3.1
requests==2.26.0
rfc3987==1.3.8
ruamel.yaml.clib==0.2.6; platform_python_implementation == 'CPython' and python_version < '3.10'
ruamel.yaml==0.17.16
shexjsg==0.7.1
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
sparqlslurper==0.4.1
sparqlwrapper==1.8.5
sphinx-click==3.0.1
sphinx-rtd-theme==1.0.0
sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlalchemy==1.4.26
strsimpy==0.2.1
tabulate==0.8.9
testfixtures==6.18.3
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchdog==2.1.6
watchgod==0.7
websockets==10.0
wrapt==1.13.3rc1
zipp==3.6.0
